refactor: new typeorm connect. refactored filesstructure.
