+++ ormconfig.json
@@ -1,11 +1,11 @@
 {
     "type": "mysql",
-    "host": "mysql",
+    "host": "localhost",
     "port": 3306,
     "username": "root",
     "password": "123698741",
     "database": "test",
-    "entities": ["dist/**/*.entity{.ts,.js}"],
+    "entities":  ["build/**/*.entity.js"],
     "synchronize": true,
     "logging": false
   }
