+++ src/account/account.controller.ts
@@ -1,13 +1,7 @@
 import { Controller, Get, Post, HttpCode, Body,  } from '@nestjs/common'
-import { AppProvider } from './app.provider'
-import { User } from './db/models/user.entity'
-import { Account } from './db/models/account.entity'
+import { AccountProvider } from './account.provider'
+import { Account } from './account.entity'
 import { ApiBody, ApiCreatedResponse, ApiProperty } from '@nestjs/swagger'
-
-export class UserBody {
-    @ApiProperty({ type: String, description: 'email' })
-    email: string;
-}
 
 export class PaymentBody {
     @ApiProperty({ type: Number, description: 'cash amount' })
@@ -29,43 +23,29 @@ export class TransferBody {
     userTo: string;
 }
 
-@Controller()
-export class AppController {
-    constructor(private readonly appProvider: AppProvider = new AppProvider) {}
+@Controller('accounts')
+export class AccountController {
+    constructor(private readonly accountProvider: AccountProvider) {}
 
-    @Get('/users')
-    @ApiCreatedResponse({ description: 'Метод отдаёт список всех Пользователей и связанных с ними аккаунтов' })
-    getUser(): Promise<User[]> {
-        return this.appProvider.getUsers();
-    }
-
-    @Get('/accounts')
+    @Get('/list')
     @ApiCreatedResponse({ description: 'Метод отдаёт список всех аккаунтов. Довольно бесполезен))' })
     getAccount(): Promise<Account[]> {
-        return this.appProvider.getAccounts();
-    }
-    
-    @Post('/user')
-    @ApiCreatedResponse({ description: 'Метод создания пользователя, Аккаунт для пользователя создаётся автоматически' })
-    @HttpCode(200)
-    @ApiBody({ type: UserBody })
-    createUser(@Body() body: UserBody): void {
-        return this.appProvider.createUser(body);
+        return this.accountProvider.getAccounts();
     }
 
-    @Post('/accounts/payment')
+    @Post('/payment')
     @ApiCreatedResponse({ description: 'Метод зачисления денежных средств на аккаунт из сторонней систему' })
     @HttpCode(200)
     @ApiBody({ type: PaymentBody })
     Payment(@Body() body: PaymentBody) {
-        return this.appProvider.makePayment(body);
+        return this.accountProvider.makePayment(body);
     }
 
-    @Post('/accounts/transfer')
+    @Post('/transfer')
     @ApiCreatedResponse({ description: 'Метод перевода денежных средств с аккаунта на аккаунт. Изначально не идемпотентен. Желательно передавать paymentId' })
     @HttpCode(200)
     @ApiBody({ type: TransferBody })
     Transfer(@Body() body: TransferBody) {
-        return this.appProvider.makeTransfer(body);
+        return this.accountProvider.makeTransfer(body);
     }
 }
