+++ src/app.model.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common'
-import { AppController } from './app.controller'
-import { AppProvider } from './app.provider'
 import { TypeOrmModule, } from '@nestjs/typeorm';
 import { getConnectionOptions, Connection } from 'typeorm'
+import { UsersModule } from './user/users.model'
+import { AccountModule } from './account/account.model'
 
 @Module({
     imports: [
@@ -11,12 +11,10 @@ import { getConnectionOptions, Connection } from 'typeorm'
             Object.assign(await getConnectionOptions(), {
               autoLoadEntities: true,
             }),
-        })
+        }),
+        UsersModule,
+        AccountModule
     ],
-    controllers: [AppController],
-    providers: [AppProvider],
 })
 
-export class AppModule {
-    constructor(private connection: Connection) {}
-}
+export class AppModule {}
